clock tree viewer: mock text report from clock tree viewer
it would be nice to have a tcl procedure to tinker with that
would print out information about the clock tree.

text is searchable, which the clock tree viewer is not.

also the clock tree viewer icons are tiny and it is hard
to hover over something to try to find it, text
is searchable.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>